\name{rotation}
\alias{rotation}
\title{Rotate 2D points}
\description{
 rotate points, counterclockwise for positive angles, and clockwise for negative ones
}
\usage{
rotation(coords, radian)
}
\arguments{
  \item{coords}{2-col matrix of [x,y] coordinates}
  \item{radian}{rotation angle}
}
\value{a 2-col matrix with the points rotated around [0,0]
}
\author{Javier Garcia-Pintado}
\keyword{ utilities }
