\name{SpatialGraph-package}
\alias{SpatialGraph-package}
\alias{SpatialGraph-package}
\docType{package}
\title{
\packageTitle{SpatialGraph}
}
\description{
\packageDescription{SpatialGraph}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SpatialGraph}
\packageIndices{SpatialGraph}
see the documentation of the function sl2sg in this package to get a
start. A case study making use if this package is Garcia-Pintado et al (2015)
}
\author{
\packageAuthor{SpatialGraph}

Maintainer: \packageMaintainer{SpatialGraph}
}
\references{
The first published application of this package is
Garcia-Pintado, J. et al. (2015). Satellite-supported flood forecasting in
river networks: a real case study. J. Hydrol. 523, 705-724.
}
\keyword{ package }

