\name{polylineLength}
\alias{polylineLength}
\title{Obtain the length of a polyline}
\description{
  Obtain the length a polyline [2-col matrix]
}
\usage{
polylineLength(xy)
}
\arguments{
  \item{xy}{a 2-column matrix representing the polyline nodes}
}
\details{
 \code{polylineLength} calculates the
 [along-polyline] length of the polyline
}
\value{
  A scalar
}
\author{Javier Garcia-Pintado, e-mail:
  \email{j.garcia-pintado@reading.ac.uk}}
\seealso{\code{\link{polylineChainage}}}
\keyword{ utilities }
\keyword{ spatial }
