\name{SpatialGraph-package}
\alias{SpatialGraph-package}
\alias{SpatialGraph-package}
\docType{package}
\title{
\packageTitle{SpatialGraph}
}
\description{
\packageDescription{SpatialGraph}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SpatialGraph}
\packageIndices{SpatialGraph}
~~ An overview of how to use the package ~~
}
\author{
\packageAuthor{SpatialGraph}

Maintainer: \packageMaintainer{SpatialGraph}
}
\references{
The first published application of this package is
Garcia-Pintado, J., Mason, D.C., Dance, S.L.,  Cloke, H.L., Neal, J.C.,
Freer. J., Bates, P.D., 2015. Satellite-supported flood forecasting in
reiver networks: a real case study. J. Hydrol. 523, 705-724.
}
\keyword{ package }

